refactor!: move pnet into separate package (#2165)
- Move pnet module to a separate package
- Related to #1913
