--- packages/libp2p/test/fixtures/swarm.key.ts
@@ -1,3 +1,0 @@
-export default '/key/swarm/psk/1.0.0/\n' +
-  '/base16/\n' +
-  '411f0a244cbbc25ecbb2b070d00a1832516ded521eb3ee3aa13189efe2e2b9a2'
