+++ packages/pnet/package.json
@@ -1,0 +1,67 @@
+{
+  "name": "@libp2p/pnet",
+  "version": "1.0.0",
+  "description": "Implementation of Connection protection management via a shared secret",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p/tree/master/packages/pnet#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p/issues"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "project": true,
+      "sourceType": "module"
+    }
+  },
+  "scripts": {
+    "build": "aegir build",
+    "test": "aegir test",
+    "clean": "aegir clean",
+    "lint": "aegir lint",
+    "test:chrome": "aegir test -t browser --cov",
+    "test:chrome-webworker": "aegir test -t webworker",
+    "test:firefox": "aegir test -t browser -- --browser firefox",
+    "test:firefox-webworker": "aegir test -t webworker -- --browser firefox",
+    "test:node": "aegir test -t node --cov",
+    "dep-check": "aegir dep-check"
+  },
+  "dependencies": {
+    "@libp2p/crypto": "^2.0.5",
+    "@libp2p/interface": "^0.1.3",
+    "@libp2p/logger": "^3.0.2",
+    "it-handshake": "^4.1.3",
+    "it-map": "^3.0.3",
+    "it-pair": "^2.0.6",
+    "it-pipe": "^3.0.1",
+    "it-stream-types": "^2.0.1",
+    "uint8arrays": "^4.0.6",
+    "xsalsa20": "^1.1.0"
+  },
+  "devDependencies": {
+    "@libp2p/interface-compliance-tests": "^4.1.1",
+    "@libp2p/peer-id-factory": "^3.0.5",
+    "@multiformats/multiaddr": "^12.1.5",
+    "@types/xsalsa20": "^1.1.0",
+    "aegir": "^41.0.2",
+    "it-all": "^3.0.1"
+  }
+}
